% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bayes.factor}
\alias{bayes.factor}
\title{TODO: Compute Bayes factor from powder objects NOT IMPLEMENTED}
\usage{
bayes.factor(m1, m2, log = TRUE)
}
\arguments{
\item{m1}{a Powder.Hierarchical or Powder.Individual object}

\item{m2}{a Powder.Hierarchical or Powder.Individual object}

\item{log}{When TRUE, the natural log of the Bayes factor will be returned.}
}
\value{
The natural log of the Bayes factor in terms of m1 / m2.
}
\description{
\code{bayes.factor} computes the Bayes factor from the powder objects
}
