% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{crossover}
\alias{crossover}
\title{Differential Evolution functions}
\usage{
crossover(i, n.chains, b, pars, use.theta, use.like, data, hyper, par.names,
  temperature, model)
}
\description{
Differential Evolution functions
}
