% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_ind.R
\docType{class}
\name{DDM.Individual}
\alias{DDM.Individual}
\title{Individual subject Drift Diffusion Model (DDM)}
\format{\code{\link{R6Class}} object.}
\usage{
DDM.Individual
}
\arguments{
\item{a}{threshold seperation}

\item{v}{drift rate}

\item{t0}{non-decision time}

\item{z}{relative starting point}

\item{st0}{inter-trial-variability of non-decision time}

\item{sv}{inter-trial-variability of drift rate}

\item{sz}{inter-trial-variability of starting point}

\item{prior}{list containing priors}
}
\description{
Individual subject Drift Diffusion Model (DDM)
}
\section{Fields}{

\describe{
\item{\code{theta.names}}{a character vector containing the names of the subject-level parameters}

\item{\code{theta.init}}{a function that provides a random initial value for each subject-level parameter}

\item{\code{theta.start.point}}{a numeric vector containing means of start points used to initialize in theta.init}

\item{\code{vary.parameter}}{a logical vector containing parameters to vary}

\item{\code{prior}}{a list containing priors on all parameters}
}}

\section{Methods}{

\describe{
\item{\code{log.dens.prior(x,hyper)}}{likelihood of subject-level parameters given group-level parameters}
\item{\code{log.dens.like(x,data,par.names)}}{LBA likelihood function}}
}

\examples{
\dontrun{
# DDM model that varies threshold and drift rate across 3 conditions
# Responses of "1" correspond to the lower response boundary and
# resonses of "2" correspond to the upper response boundary.
model = DDM.Individual$new(a=T,v=T,conds=1:3)

# Note, inter-trial-variability components, st0, sv, sz, and
# start point, z, have 3 options: TRUE, FALSE, or 0. When TRUE,
# they will vary across conditions. When FALSE
# they are fit, but not allowed to vary across conditions.
# By default, they are set to 0.
# Note, when z = 0 this indicates no bias (z is internally set to .5).

# st0 is a parameter, other variability parameters are 0
model_st0 = DDM.Individual$new(st0=FALSE)

# sv is allowed to vary over conditions
model_sv = DDM.Individual$new(sv=TRUE)

# by default z is internally set to .5
model_v = DDM.Individual$new(v=TRUE)
# the model above by default sets z to 0, which
# indicates that z will be set to .5 internally.


 # Setting priors:
 prior = list(
 a = c(mu=1,sigma=1),
 v = c(mu=1,sigma=1),
 t0 = c(mu=.3,sigma=.3),
 z = c(mu=.5,sigma=.5),
 sz = c(mu=.2,sigma=.2),
 sv = c(mu=.1,sigma=.1))

 # a and v vary across conditions
 # sz and sv are parameters and do not vary across conditions
 # st0 is 0
 model = DDM.Individual$new(a=TRUE,v=TRUE,sz=FALSE,sv=FALSE)

 # a and v vary across conditions
 # sz and sv are parameters and do not vary across conditions
 # st0 is 0
 # z is a fit parameter but does not vary across conditions
 model = DDM.Individual$new(a=TRUE,v=TRUE,sz=FALSE,sv=FALSE,z=FALSE)

 # Priors can also be changed after creating the model
 model$prior$a = c(mu=2,sigma=2)

}
}
\keyword{datasets}
