% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\docType{class}
\name{DDM}
\alias{DDM}
\title{Individual subject Drift Diffusion Model (DDM)}
\format{\code{\link{R6Class}} object.}
\usage{
DDM
}
\arguments{
\item{a}{threshold seperation}

\item{v}{drift rate}

\item{t0}{non-decision time}

\item{z}{starting point}

\item{st0}{inter-trial-variability of non-decision time}

\item{sv}{inter-trial-variability of drift rate}

\item{sz}{inter-trial-variability of starting point}

\item{prior}{list containing priors}

\item{conds}{numeric vector containing values corresponding to conditions}

\item{correct.response}{a character vector. Specifies the value corresponding to the correct response}

\item{error.response}{a character vector. Specifies the value corresponding to the error response}

\item{rt.column}{a character vector. Specifies the name of the column in the data for response times.}

\item{response.column}{a character vector. Specifies the name of the column in the data for respones.}

\item{condition.column}{a character vector. Specifies the name of the column in the data for the conditions.}
}
\description{
Individual subject Drift Diffusion Model (DDM)
}
\section{Fields}{

\describe{
\item{\code{theta.names}}{a character vector containing the names of the subject-level parameters}

\item{\code{theta.init}}{a function that provides a random initial value for each subject-level parameter}

\item{\code{theta.start.point}}{a numeric vector containing means of start points used to initialize in theta.init}

\item{\code{vary.parameter}}{a logical vector containing parameters to vary}

\item{\code{prior}}{a list containing priors on all parameters}
}}

\section{Methods}{

\describe{
\item{\code{log.dens.prior(x,hyper)}}{likelihood of subject-level parameters given group-level parameters}
\item{\code{log.dens.like(x,data,par.names)}}{LBA likelihood function}}
}

\examples{
\dontrun{
# DDM model that varies threshold and drift rate across 3 conditions
model = DDM.Individual$new(a=T,v=T,conds=1:3)

# Note, inter-trial-variability components, st0, sv, and sz
# have 3 options: TRUE, FALSE, or 0. When TRUE,
# they will vary across conditions. When FALSE
# they are fit, but not allowed to vary across conditions.
By default, they are set to 0.

# st0 is a parameter, other variability parameters are 0
model_st0 = DDM$new(st0=FALSE)

# sv is allowed to vary over conditions
model_sv = DDM$new(sv=TRUE)

# Setting priors:
prior = list(
    a = list(mu = c(mu = 1, sigma = 1),
             sigma = c(mu = 1, sigma = 1)),
    v = list(mu = c(mu = 1, sigma = 1),
             sigma = c(mu = 1, sigma = 1)),
    t0 = list(
         mu = c(mu = .3, sigma = .3),
         sigma = c(mu = .3, sigma = .3)),
    st0 = list(
         mu = c(mu = .1, sigma = .1),
         sigma = c(mu = .1, sigma = .1)))

 # a and v vary across conditions
 # sz and sv are 0
 # t0 and st0 are parameters that do not vary across conditions
 model = DDM$new(a=TRUE,v=TRUE,prior=prior)

 # Priors can also be changed after creating the model
 model$prior$a$mu = c(mu=2,sigma=2)

}
}
\keyword{datasets}
