% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drift}
\alias{drift}
\title{Simulated example dataset from the LBA (varies drift rate across conditions)}
\format{A list with 10 elements, each representing a simulated subject. Each element contains the following:
\describe{
\item{Cond}{condition number}
\item{Time}{response time}
\item{Correct}{whether the response was error (2) or correct (1)}
}}
\usage{
drift
}
\description{
A simulated dataset from the LBA with 10 subjects, 2 conditions, and 300 trials per condition.
Simulates correct mean drift rate varying across conditions.
The parameters used to generate the null dataset were the following:
\itemize{
\item starting point variability, A: 1
\item relative threshold, b: 0.4
\item correct drift rate mean condition 1, vc.1: 2
\item correct drift rate mean condition 2, vc.2: 3.5
\item error drift rate mean, ve: 1
\item drift rate variability, svc, sve: 1
\item non-decision time, t0: 0.3
}
}
\keyword{datasets}
