% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individual}
\alias{individual}
\title{The "Simple" dataset from Evans and Brown (2017)}
\format{A list with 3 elements:
\describe{
  \item{Cond}{condition number}
  \item{Correct}{whether the response was error (2) or correct (1)}
  \item{Time}{response time}
}}
\usage{
individual
}
\description{
The "Simple" dataset from Evans and Brown (2017)
}
\keyword{datasets}
