% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lba.R
\docType{class}
\name{LBA}
\alias{LBA}
\title{The Linear Ballistic Accumulator model (Brown & Heathcote, 2008)}
\format{\code{\link{R6Class}} object.}
\usage{
LBA
}
\value{
Object of \code{\link{R6Class}} with methods for doing sampling with \code{\link{powder}}
}
\description{
The Linear Ballistic Accumulator model (Brown & Heathcote, 2008)
}
\section{Fields}{

\describe{
\item{\code{theta.names}}{a character vector containing the names of the subject-level parameters}

\item{\code{theta.init}}{a function that provides a random initial value for each subject-level parameter}

\item{\code{theta.start.point}}{a numeric vector containing means of start points used to initialize in theta.init}

\item{\code{phi.names}}{a character vector containing the names of the group-level parameters}

\item{\code{phi.init}}{a function that initializes the group-level parameters}

\item{\code{vary.parameter}}{a logical vector containing parameters to vary}

\item{\code{prior}}{a list containing priors on all parameters}
}}

\section{Methods}{

\describe{
\item{\code{log.dens.prior(x,hyper)}}{likelihood of subject-level parameters given group-level parameters}
\item{\code{log.dens.hyper(theta,phi,prior)}}{summed log-likelihood for subject given group-level parameters and group given priors}
\item{\code{log.dens.like(x,data,par.names)}}{LBA likelihood function}}
}

\examples{
\dontrun{
#LBA model that varies threshold across 3 conditions
model_threshold = LBA$new(b=T,conds=1:3)
#LBA model that varies correct drift rate over 2 conditions
model_drift = LBA$new(vc=T,conds=1:2)
#LBA model that varies start point variability and non-decision over 2 conditions
model_sp_t0 = LBA$new(t0=T,A=T,conds=1:2)
}
}
\keyword{datasets}
