% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{load_rdata}
\alias{load_rdata}
\title{Load an RData file}
\usage{
load_rdata(file_name)
}
\arguments{
\item{file_name}{a character string giving the name of the file to load}
}
\value{
The objects in the RData file. If there is more than one object, a list will be returned.
}
\description{
\code{load_rdata} loads an RData file into a specified variable
}
\examples{
\dontrun{
x = pi # to ensure there is some data
x_list = list(pi = pi, pi_sq = pi^2)
save(x, file= "pi.RData")
save(x_list, file= "pi_list.RData")
rm(x)
rm(x_list)
x = load_rdata('pi.RData')
x_list = load_rdata('pi_list.RData')
}
}
\keyword{internal}
