% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{marginal.likelihood}
\alias{marginal.likelihood}
\title{Estimates marginal likelihood from power posterior samples}
\usage{
marginal.likelihood(pow.out)
}
\arguments{
\item{pow.out}{The output from the \code{powder} function}
}
\value{
A data frame with the following columns

\code{TI} Log marginal likelihood estimate based on thermodynamic integration

\code{TI Corrected} Log marginal likelihood estimate based on modified thermodynamic integration using variance of log likelihood

\code{Harmonic Mean} Log marginal likelihood estimate based on harmonic mean

\code{Steppingstone} Marginal likelihood esimate based on steppingstone estimator. Because of underflow, this will usually not be meaningful.

\code{Log Steppingstone} Log marginal likelihood estimate based on steppingstone estimator.

\code{TI Variance} Estimated variance of \code{TI}

\code{Steppingstone Variance} Estimated variance of \code{Log Steppingstone}
}
\description{
\code{marginal.likelihood} Estimates marginal likelihood using output from \code{powder}
}
\examples{
\dontrun{
model = LBA$new(b=T)
data('null',package='powder')
num.temps = 30
pow_out = powder(data=null, model=model, num.temps=num.temps)
}
}
