% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{null}
\alias{null}
\title{The "null" simulated dataset}
\format{A list with 10 elements, each representing a simulated subject. Each element contains the following:
\describe{
\item{Cond}{condition number}
\item{Time}{response time}
\item{Correct}{whether the response was error (2) or correct (1)}
}}
\usage{
null
}
\description{
A simulated dataset from the LBA with 10 subjects, 2 conditions, and 300 trials per condition.
Simulates no parameters varying across conditions.
The parameters used to generate the null dataset were the following:
\itemize{
\item starting point variability, A: 1
\item relative threshold, b: 0.4
\item correct drift rate mean vc: 3
\item error drift rate mean, ve: 1
\item drift rate variability, svc, sve: 1
\item non-decision time, t0: 0.3
}
}
\keyword{datasets}
