% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.Powder.Individual}
\alias{plot.Powder.Individual}
\title{Plots the samples by iteration of each chain}
\usage{
\method{plot}{Powder.Individual}(x, pars = NULL, type = "iteration",
  options = list(), ...)
}
\arguments{
\item{x}{a Powder.Individual object}

\item{pars}{a character vector of parameters to plot}

\item{type}{a character vector indicating the type of plot. Two plots are supported: `density` and `iteration`.}

\item{options}{a list that can supply any of the following components
\describe{
\item{\code{with.burnin}}{a \code{\link{logical}} indicating whether to include burnin with plot. Defaults to false.}
\item{\code{ggplot_theme}}{a ggplot theme}
\item{\code{return.plot}}{a \code{\link{logical}} indicating whether to return the ggplot object. A plot will be displayed regardless. Defaults to false.}
}}
}
\value{
A plot of the samples as a function of iteration and parameter
}
\description{
\code{plot.Powder.Individual}
}
