% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_powder.R
\name{pwr_grid}
\alias{pwr_grid}
\alias{pwr_tidy}
\alias{pwr_plot}
\title{powder: A Tidy Extension for Power Analysis}
\usage{
pwr_grid(k = NULL, n = NULL, n1 = NULL, n2 = NULL, N = NULL,
  h = NULL, f = NULL, f2 = NULL, r = NULL, w = NULL,
  diff = NULL, sigma = NULL, d = NULL, df = NULL, u = NULL,
  v = NULL, sig.level = 0.05, power = NULL, type = NULL,
  alternative = NULL)

pwr_tidy(pwr_grid, test_function)

pwr_plot(pwr_tidy, x, y, group)
}
\arguments{
\item{n}{Number of observations (per sample)}

\item{diff}{Difference between the means}

\item{sigma}{Standard deviation}

\item{d}{Effect size (Cohen's d)}

\item{sig.level}{Significance level (Type I error probability)}

\item{power}{Power of test (1 minus Type II error probability)}

\item{type}{Type of t test : one- two- or paired-samples}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}

\item{pwr_grid}{outcome of pwr_grid function}

\item{test_function}{function to calculate power from pwr or stats package}

\item{pwr_tidy}{outcome of pwr_tidy function}

\item{x}{x axis}

\item{y}{y axis}

\item{group}{colour variable}
}
\value{
canal endemico, union y grafico
}
\description{
Create tables and plots for power and sample size calculations from a range of required parameters. All of the parameters accepts a range or sequence of numbers. Usable with pwr and stats::power.
}
\section{Functions}{
\itemize{
\item \code{pwr_grid}: 

\item \code{pwr_tidy}: create table with parameters and calculations

\item \code{pwr_plot}: create diagnostic plot
}}

\examples{

library(tidyverse)
library(pwr)
library(powder)

# stata
# power onemean 20, diff(10 20 30) sd(10 12.5 15)
eg1 <- pwr_grid(n = NULL,
                diff = c(10,20,30),
                sigma = c(10,12.5,15),
                d = NULL,
                sig.level = 0.05,
                power = 0.8,
                type = "one.sample",
                alternative = "two.sided")

eg1 \%>\% pwr_tidy(test_function = pwr.t.test)

eg1 \%>\%
  pwr_tidy(test_function = pwr.t.test) \%>\%
  pwr_plot(x = diff,y = n,group = sigma)

# stata
# power onemean 20, diff(5 (5) 15) sd(10) n(10 (2) 20)
eg2 <- pwr_grid(n = seq(from = 10,to = 20,by = 2),
                diff = seq(from = 5,to = 15,by = 5),
                sigma = 10,
                d = NULL,
                sig.level = 0.05,
                #power = 0.8,
                type = "one.sample",
                alternative = "two.sided")

eg2 \%>\% pwr_tidy(test_function = pwr.t.test)

eg2 \%>\%
  pwr_tidy(test_function = pwr.t.test) \%>\%
  pwr_plot(x = n,y = power,group=diff)

# stata
# power onemean 100, n(10 20 30) power(0.8) sd(10 12.5 15)
eg3 <- pwr_grid(n=c(10,20,30),
                #diff=c(10,20,30),
                sigma=c(10,12.5,15),
                d=NULL,
                sig.level=0.05,
                power=0.8,
                type="one.sample",
                alternative="two.sided")

eg3 \%>\% pwr_tidy(test_function = power.t.test)

eg3 \%>\%
  pwr_tidy(test_function = power.t.test) \%>\%
  pwr_plot(x = n,y = delta,group=sd)

}
