% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rae}
\alias{rae}
\title{The dataset from Rae et al. (2014)}
\format{A list with 34 elements, each representing a subject. Each element contains the following:
\describe{
  \item{Cond}{condition number}
  \item{Correct}{whether the response was error (2) or correct (1)}
  \item{Time}{response time}
}}
\usage{
rae
}
\description{
The dataset from Rae et al. (2014)
}
\keyword{datasets}
