% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reshape.phi}
\alias{reshape.phi}
\title{Convert phi array to dataframe}
\usage{
reshape.phi(pow.out, pars = NULL, with.burnin = FALSE)
}
\arguments{
\item{pow.out}{Powder.Hierarchical object}

\item{pars}{subset of parameters to reshape. Defaults to all parameters.}

\item{with.burnin}{if false, burnin is discarded}
}
\value{
A dataframe with columns corresponding to parameter, iteration, chain, and value
}
\description{
Reshapes the group-level parameters in Powder.Individual or Powder.Hierarchical objects to a dataframe for easier analysis.
}
\examples{
\dontrun{
data('null')
lba = LBA$new()
#low number of samples for illustration purposes
out = powder(model=lba,data=null,burnin=10,n.samples=10,method='posterior')
phi_df = reshape.phi(out)
}
}
