% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reshape.theta}
\alias{reshape.theta}
\title{Convert theta array to dataframe}
\usage{
reshape.theta(pow.out, pars = NULL, with.burnin = FALSE, ...)
}
\arguments{
\item{pow.out}{Powder.Individual or Powder.Hierarchical object}

\item{pars}{subset of parameters to reshape. Defaults to all parameters.}

\item{with.burnin}{if false, burnin is discarded}

\item{...}{additional arguments}

\item{subject}{subset of subjects to reshape. Defaults to all subjects.}
}
\value{
A dataframe with columns corresponding to parameter, iteration, chain, value,
and subject if Powder.Hierarchical object.
}
\description{
Reshapes the subject-level parameters in Powder.Individual or Powder.Hierarchical objects to a dataframe for easier analysis.
}
\examples{
\dontrun{
data('null')
lba = LBA$new()
#low number of samples for illustration purposes
out = powder(model=lba,data=null,burnin=10,n.samples=10,method='posterior')
phi_df = reshape.phi(out)
}
}
