% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{summary.Powder.Hierarchical}
\alias{summary.Powder.Hierarchical}
\title{Estimates the marginal likelihood and summarizes the samples}
\usage{
\method{summary}{Powder.Hierarchical}(object, options = list(), ...)
}
\arguments{
\item{object}{a Powder.Hierarchical object from \code{\link{powder}}}
}
\value{
A \code{\link{tibble}}. If the Powder.* object was generated via the \code{standard} or \code{parallel} method from \code{\link{powder}} then the returned value will be a \code{\link{tibble}} with the following rows:
\describe{
\item{\code{TI}}{log marginal likelihood estimate based on thermodynamic integration}
\item{\code{TI Corrected}}{log marginal likelihood estimate based on modified thermodynamic integration using variance of log likelihood}
\item{\code{Harmonic Mean}}{log marginal likelihood estimate based on harmonic mean}
\item{\code{Steppingstone}}{marginal likelihood esimate based on steppingstone estimator. Because of underflow, this will usually not be meaningful.}
\item{\code{Log Steppingstone}}{log marginal likelihood estimate based on steppingstone estimator.}
\item{\code{TI Variance}}{estimated variance of \code{TI}}
\item{\code{Steppingstone Variance}}{estimated variance of \code{Log Steppingstone}}
}

If the object was generated via the \code{posterior} method from \code{\link{powder}} then the returned value will be a \code{\link{tibble}} with the following columns:
\describe{
\item{\code{mean}}{the mean of the parameter}
\item{\code{95\% HDI UB}}{upper bound of 95\% highest density interval}
\item{\code{95\% HDI LB}}{lower bound of 95\% highest density interval}
}
}
\description{
Summarizes the samples or estimates the marginal likelihood depending on the type of samples
}
\examples{
\dontrun{
model = LBA$new()
data('null',package='powder')
#low samples for illustration purposes only
pow.out = powder(data=null, model=model, method='parallel', burnin=10, n.samples=20)
summary(pow.out,options=list(sig.digits=6))

# # A tibble: 7 x 2
# Method                       Value
# <fct>                        <dbl>
# 1 TI                       1822.12
# 2 TI Corrected             1920.92
# 3 Harmonic Mean            2128.10
# 4 Steppingstone            NaN
# 5 Log Steppingstone        2086.49
# 6 TI Variance              4144.37
# 7 Steppingstone Variance   0.58262

pow.out = powder(data=null, model=model, method='posterior',burnin=10,n.samples=20)
summary(pow.out,options=list(sig.digits=3,num.rows=20))

# $group.level
# # A tibble: 12 x 4
# parameter  mean         95\% HDI LB         95\% HDI UB
# <fct>     <dbl>              <dbl>               <dbl>
# 1 A.mu      0.769            0.0803               1.33
# 2 A.sigma   0.781            0.275                1.25
# 3 b.mu      0.750            0.0459               1.29
# 4 b.sigma   0.724            0.177                1.46
# 5 sve.mu    0.924            0.115                1.41
# 6 sve.sigma 0.687            0.242                1.23
# 7 t0.mu     0.150            0.00687              0.227
# 8 t0.sigma  0.122            0.0488               0.240
# 9 vc.mu     3.22             0.427                4.61
# 10 vc.sigma 1.84             0.731                3.98
# 11 ve.mu    1.57             0.262                3.10
# 12 ve.sigma 1.49             0.278                2.97
#
# $subject.level
# # A tibble: 60 x 5
# parameter subject  mean   95\% HDI LB    95\% HDI UB
# <fct>     <fct>   <dbl>        <dbl>         <dbl>
# 1 A         1       0.850      0.194          1.61
# 2 b         1       1.14       0.125          1.89
# 3 sve       1       1.05       0.549          1.49
# 4 t0        1       0.141      0.00486        0.271
# 5 vc        1       3.61       2.42           4.59
# 6 ve        1       1.60       0.238          2.84
# 7 A         10      0.943      0.232          1.68
# 8 b         10      0.950      0.340          1.44
# 9 sve       10      1.27       0.866          1.70
# 10 t0       10      0.205      0.0877         0.321
# # ... with 50 more rows
}
}
