% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{summary.Powder.Individual}
\alias{summary.Powder.Individual}
\title{Estimates marginal likelihood from power posterior samples}
\usage{
\method{summary}{Powder.Individual}(object, options = list(), ...)
}
\arguments{
\item{object}{A Powder.Individual or Powder.Hierarchical object from \code{\link{powder}}}
}
\value{
A \code{\link{tibble}}. If the Powder.* object was generated via the \code{standard} or \code{parallel} method from \code{\link{powder}} then the returned value will be a \code{\link{tibble}} with the following rows:
\describe{
\item{\code{TI}}{log marginal likelihood estimate based on thermodynamic integration}
\item{\code{TI Corrected}}{log marginal likelihood estimate based on modified thermodynamic integration using variance of log likelihood}
\item{\code{Harmonic Mean}}{log marginal likelihood estimate based on harmonic mean}
\item{\code{Steppingstone}}{marginal likelihood esimate based on steppingstone estimator. Because of underflow, this will usually not be meaningful.}
\item{\code{Log Steppingstone}}{log marginal likelihood estimate based on steppingstone estimator.}
\item{\code{TI Variance}}{estimated variance of \code{TI}}
\item{\code{Steppingstone Variance}}{estimated variance of \code{Log Steppingstone}}
}

If the object was generated via the \code{posterior} method from \code{\link{powder}} then the returned value will be a \code{\link{tibble}} with the following columns:
\describe{
\item{\code{mean}}{the mean of the parameter}
\item{\code{95\% HDI UB}}{upper bound of 95\% highest density interval}
\item{\code{95\% HDI LB}}{lower bound of 95\% highest density interval}
}
}
\description{
Estimates marginal likelihood using output from \code{powder}
}
\examples{
\dontrun{
model = LBA.Individual$new()
data('individual',package='powder')
# low number of samples for illustration purposes only
pow.out = powder(data=individual, model=model, method='parallel',burnin=100,n.samples=500)
summary(pow.out,options=list(sig.digits=5))

# A tibble: 7 x 2
# Method                       Value
# <fct>                        <dbl>
# 1 TI                     5.6759e+  2
# 2 TI Corrected           5.6799e+  2
# 3 Harmonic Mean          5.6184e+  2
# 4 Steppingstone          6.1592e+245
# 5 Log Steppingstone      5.6955e+  2
# 6 TI Variance            3.6981e-  2
# 7 Steppingstone Variance 9.8905e-  2

# do posterior estimation
pow.out = powder(data=individual, model=model, method='posterior',burnin=100,n.samples=500)
summary(pow.out,options=list(sig.digits=3))

#   parameter  mean   95\% HDI LB   95\% HDI UB
#   <fct>     <dbl>        <dbl>        <dbl>
# 1 A         0.827        0.553        1.02
# 2 b         0.575        0.261        1.01
# 3 sve       1.18         0.876        1.46
# 4 t0        0.269        0.178        0.320
# 5 vc        3.44         2.95         3.90
# 6 ve        1.12         0.655        1.60
}
}
