% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lba.R
\docType{class}
\name{LBA}
\alias{LBA}
\title{Hierarchical Linear Ballistic Accumulator model (Brown & Heathcote, 2008)}
\format{\code{\link{R6Class}} object.}
\usage{
LBA
}
\arguments{
\item{A}{threshold seperation}

\item{b}{threshold}

\item{vc}{correct drift rate}

\item{ve}{error drift rate}

\item{t0}{non-decision time}

\item{sve}{inter-trial-variability of error drift rate. Note that correct drift rate variability is fixed at 1.}

\item{prior}{list containing priors}

\item{conds}{numeric vector containing values corresponding to conditions}

\item{correct.response}{a character vector. Specifies the value corresponding to the correct response}

\item{error.response}{a character vector. Specifies the value corresponding to the error response}

\item{rt.column}{a character vector. Specifies the name of the column in the data for response times.}

\item{response.column}{a character vector. Specifies the name of the column in the data for respones.}

\item{condition.column}{a character vector. Specifies the name of the column in the data for the conditions.}

\item{contaminant}{a list specifying two values
\describe{
\item{\code{pct}}{the percentage of the LBA distribution assumed to be due to random contaminants.}
\item{\code{contaminant_bound}}{the upper bound of the contaminant distribution.}
The contaminant distribution is assumed to be a uniform spanning from 0 to \code{contaminant_bound}.
}}
}
\value{
Object of \code{\link{R6Class}} with methods for doing sampling with \code{\link{powder}}
}
\description{
Hierarchical Linear Ballistic Accumulator model (Brown & Heathcote, 2008)
}
\section{Fields}{

\describe{
\item{\code{theta.names}}{a character vector containing the names of the subject-level parameters}

\item{\code{theta.init}}{a function that provides a random initial value for each subject-level parameter}

\item{\code{theta.start.point}}{a numeric vector containing means of start points used to initialize in theta.init}

\item{\code{phi.names}}{a character vector containing the names of the group-level parameters}

\item{\code{phi.init}}{a function that initializes the group-level parameters}

\item{\code{vary.parameter}}{a logical vector containing parameters to vary}

\item{\code{prior}}{a list containing priors on all parameters}

\item{\code{contaminant}}{a list specifying two values
\describe{
\item{\code{pct}}{the percentage of the LBA distribution assumed to be due to random contaminants.}
\item{\code{contaminant_bound}}{the upper bound of the contaminant distribution.}
The contaminant distribution is assumed to be a uniform spanning from 0 to \code{contaminant_bound}.
}}
}}

\section{Methods}{

\describe{
\item{\code{log.dens.prior(x,hyper)}}{likelihood of subject-level parameters given group-level parameters}
\item{\code{log.dens.hyper(theta,phi,prior)}}{summed log-likelihood for subject given group-level parameters and group given priors}
\item{\code{log.dens.like(x,data,par.names)}}{LBA likelihood function}}
}

\examples{
\dontrun{
#LBA model that varies threshold across 3 conditions
model_threshold = LBA$new(b=T,conds=1:3)
#LBA model that varies correct drift rate over 2 conditions
model_drift = LBA$new(vc=T,conds=1:2)
#LBA model that varies start point variability and non-decision over 2 conditions
model_sp_t0 = LBA$new(t0=T,A=T,conds=1:2)
}
}
\keyword{datasets}
