% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{traceplot}
\alias{traceplot}
\title{Plots the samples by iteration of each chain}
\usage{
traceplot(x, pars = NULL, options = list(), ...)
}
\arguments{
\item{x}{a Powder.Individual or Powder.Hierarchical object}

\item{pars}{a character vector of parameters to plot}

\item{options}{a list that can supply any of the following components
\describe{
\item{\code{with.burnin}}{a \code{\link{logical}} indicating whether to include burnin with plot. Defaults to false.}
\item{\code{ggplot_theme}}{a ggplot theme}
\item{\code{return.plot}}{a \code{\link{logical}} indicating whether to return the ggplot object. A plot will be displayed regardless. Defaults to false.}
}}

\item{...}{additional arguments}

\item{subject}{a numeric vector of subject numbers to plot. If not supplied, the group-level parameters will be plotted.}
}
\value{
A plot of the samples as a function of iteration, parameter, and if supplied, subject
}
\description{
\code{traceplot}
}
